requirement.txt
streamlit
pandas
folium
streamlit-folium
requests
python-dotenv
scikit-learn